Refactoring Meraki code
